% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_random}
\alias{datos_random}
\title{Conjunto de datos random}
\format{
Los datos son generados de manera aleatoria
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por 5}
}
}
\source{
Fuente: generacion propia con funciones de R.
}
\usage{
datos_random
}
\description{
Los datos son utilizados para chequear nuestra funcion \code{suma()}.
}
\keyword{datasets}
