% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de 2 numeros}
\usage{
suma(x = 5, y = 2)
}
\arguments{
\item{x}{vector numerico}

\item{y}{vector numerico}
}
\value{
el resultado de una suma entre 2 numeros
}
\description{
Suma de 2 numeros
}
\examples{
suma(2,2)
}
